+++ NEWS.md
@@ -1,3 +1,9 @@
+0.1.5.9001: 
+
+- `plot.bamdata` and `plot.bamval` have been moved from `plot` to a new generic, `bam_plot`. 
+- Documentation has been updated based on notes in `R CMD check`. 
+- NSE-referenced objects now use `rlang::.data` to avoid check complaints. 
+
 0.1.3: 
 
 - Support missing data in observations